Separate parsing and command tests in separate files
