+++ tests/functional/parsing/test_parsing_init_only.py
@@ -1,0 +1,36 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+"""Perform functional tests for the parsing of `init_only` command."""
+import os
+
+import numpy
+import pytest
+import argparse
+
+from orion.core.io.database import Database
+from orion.core.cli import init_only
+
+def create_parser(need_subparser = True):
+    parser = argparse.ArgumentParser()
+
+    if need_subparser:
+        subparsers = parser.add_subparsers()
+        return parser, subparsers
+    
+    return parser
+
+@pytest.mark.usefixtures("clean_db")
+def test_init_only_command_full_parsing(database, monkeypatch):
+    monkeypatch.chdir(os.path.dirname(os.path.abspath(__file__)))
+    parser, subparsers = create_parser()
+    args_list = ["init_only", "-n", "test", "--config", "./orion_config_random.yaml", "./black_box.py", "-x~normal(1,1)"]
+    
+    init_only.get_parser(subparsers)
+    subparsers.choices['init_only'].set_defaults(func='')
+
+    args = vars(parser.parse_args(args_list))
+    assert args['name'] == 'test'
+    assert args['config'].name == './orion_config_random.yaml'
+    assert args['user_script'] == './black_box.py'
+    assert args['user_args'] == ['-x~normal(1,1)']
+
